python/requirements.txt
Pillow
cairosvg
numpy
spidev
opencv-python-headless
Flask
openai
requests
sounddevice
python-dotenv
